% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{module}
\alias{module}
\title{Linux Environment Modules interface}
\usage{
module(Arguments)
}
\arguments{
\item{Arguments}{"[ switches ] [ subcommand ] [ subcommand-args ]" See examples.}
}
\value{
Output messages will be sent to stderr
}
\description{
Access the linux Environment Modules which provides for the dynamic modification of a user's
environment via modulefiles.
}
\details{
Each modulefile contains the information needed to configure the shell for an application.
Once the RLinuxModules package is initialized (with \code{\link{moduleInit}}), the environment can
be modified on a per-module basis using the module command which interprets modulefiles. Typically
modulefiles instruct the module command to alter or set shell environment variables such as PATH,
MANPATH, etc. modulefiles may be shared by many users on a system and users may have their own
collection to supplement or replace the shared modulefiles.
}
\examples{
\dontrun{
module("avail") # shows available modules

module("--help") # show available sub-commands and switches

module("load samtools") # loads the module "samtools"
system("which samtools") # check that samtools is loaded in the environment

module("list") # list loaded modules

module("unload samtools") # unload the samtools module

module("load samtools/1.0") # loads a specific version of the module "samtools"
system("which samtools") # check that the correct samtools is loaded in the environment
}
}
