% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleInit.R
\name{moduleInit}
\alias{moduleInit}
\title{Initialize Environment Modules interface}
\usage{
moduleInit(version = "3.2.10",
  modulesHome = "/local/genome/Modules/3.2.10")
}
\arguments{
\item{version}{The version of the module system that is installed}

\item{modulesHome}{Path to where the module system is installed}
}
\description{
Initialize linux Environment Modules. Must be called before using \code{\link{module}}
}
